Problem 2 solved
The second scenario tests the functionality for clicking a "Find Movies
With Same Director" link on the movie-details page for a movie that has
director info in the database. For this scenario, you'll have to modify
the existing "Show Movie" view, and you'll have to add a route, view
and controller method for "find movies with same director".
- a RESTful route for "find similar movies" (Hint: Use the 'match'
syntax for routes as suggested in "Non-Resource-Based Routes" at the
end of section 3.9 of ELLS),
- a controller method to receive the click on "Find Movies With Same
Director", and grab the id (for example) of the current movie (i.e. the
one we're trying to find movies similar to), and
- a model method in the Movie model to find movies whose director
matches that of the current movie.

The back to cucumber